+++ src/manifolds/Elliptope.jl
@@ -85,10 +85,16 @@ zero diagonal.
 The tolerance for the base point check and zero diagonal can be set using the `kwargs...`.
 Note that symmetric of $X$ holds by construction an is not explicitly checked.
 """
-function check_vector(M::Elliptope, q, Y; kwargs...)
+function check_vector(
+    M::Elliptope,
+    q,
+    Y::T;
+    atol::Real=sqrt(prod(representation_size(M))) * eps(real(float(number_eltype(T)))),
+    kwargs...,
+) where {T}
     X = q * Y' + Y * q'
     n = diag(X)
-    if !all(isapprox.(n, 0.0; kwargs...))
+    if !all(isapprox.(n, 0.0; atol=atol, kwargs...))
         return DomainError(
             n,
             "The vector $(X) is not a tangent to a point on $(M) (represented py $(q) and $(Y), since its diagonal is nonzero.",