+++ test/manifolds/circle.jl
@@ -288,4 +288,13 @@ using Manifolds: TFVector, CoTFVector
             )
         end
     end
+    @testset "Mixed array dimensions for exp and PT" begin
+        # this is an issue on Julia 1.6 but not later releases
+        M = Circle()
+        p = fill(0.0)
+        exp!(M, p, p, [1.0], 2.0)
+        @test p ≈ fill(2.0)
+        parallel_transport_to!(M, p, p, [4.0], p)
+        @test p ≈ fill(4.0)
+    end
 end